refactor: reorg. of commands and subcommands
see #26
