+++ cmd/kconf/ctx/rm.go
@@ -1,40 +1,31 @@
-package clustermod
+package ctx
 
 import (
 	"github.com/sn3d/kconf/pkg/kconf"
 	"github.com/urfave/cli/v2"
 )
 
-var Cmd = &cli.Command{
-	Name:      "clustermod",
-	Usage:     "modify a cluster",
-	ArgsUsage: "[CLUSTER]",
+var rmCmd = &cli.Command{
+	Name:      "rm",
+	Usage:     "remove given context",
+	ArgsUsage: "[CONTEXT]",
 	Flags: []cli.Flag{
 		&cli.StringFlag{
 			Name:  "kubeconfig",
 			Usage: "path to kubeconfig from where export context",
 		},
-		&cli.StringFlag{
-			Name:  "url",
-			Usage: "set the cluster URL",
-		},
 	},
 
 	// main entry point for 'export'
 	Action: func(cCtx *cli.Context) error {
+		contextName := cCtx.Args().First()
+
 		kc, path, err := kconf.Open(cCtx.String("kubeconfig"))
 		if err != nil {
 			return err
 		}
 
-		opts := &kconf.ClustermodOptions{
-			ServerURL: cCtx.String("url"),
-		}
-
-		err = kc.Clustermod(cCtx.Args().First(), opts)
-		if err != nil {
-			return err
-		}
+		kc.Remove(contextName)
 
 		err = kc.Save(path)
 		if err != nil {
